Extracted Microsoft PPLX out of C++ REST SDK v2.9.1.
[Problem description]
A while back I needed more or less mature means
for task parallelism in C++, featuring some (at
least basic) facilities for task scheduling and
composition. Microsoft PPL happened to be quite
close to what I was after, but it wasn't ported
to *nix by the time (for what it's worth, it is
still not ported as of this writing). Microsoft
PPLX, a trimmed-down version of PPL, on the
other hand, was cross platform, but it came
with its own pain: was bundled within a library
called C++ REST SDK, Microsoft's offering for
creating RESTful services in C++. Hence you had
to include C++ REST SDK in its entirety to have
basic means for generic concurrency.

[Problem solution]
Pulled PPLX out of C++ REST SDK v2.9.1, presen-
ting it in a form of a separate shared or sta-
tic library.